refactor(config): extract bootstrap source modules
