+++ runtime/src/core/config-sources.ts
@@ -1,0 +1,153 @@
+import { readEnvFile } from "./env.js";
+
+/** Piclaw names read from `.env` before typed domain resolution. */
+export const ENV_FILE_KEYS = [
+  "PICLAW_ASSISTANT_NAME",
+  "PICLAW_ASSISTANT_AVATAR",
+  "PICLAW_USER_NAME",
+  "PICLAW_USER_AVATAR",
+  "PICLAW_USER_AVATAR_BACKGROUND",
+  "ASSISTANT_NAME",
+  "ASSISTANT_AVATAR",
+  "PICLAW_AGENT_TIMEOUT",
+  "AGENT_TIMEOUT",
+  "PICLAW_BACKGROUND_AGENT_TIMEOUT",
+  "AGENT_TIMEOUT_BACKGROUND",
+  "PICLAW_DREAM_CRON",
+  "PICLAW_DREAM_BACKUP_KEEP",
+  "PICLAW_DREAM_MODEL",
+  "PICLAW_DREAM_AGENT_TIMEOUT_MS",
+  "PICLAW_WEB_MAX_CONTENT_CHARS",
+  "PICLAW_WEB_PREVIEW_CHARS",
+  "PICLAW_RECORDINGS_DIR",
+  "PICLAW_ADDON_API_FAILURE_BACKOFF_MS",
+  "PICLAW_ABORT_SETTLE_TIMEOUT_MS",
+  "PUSHOVER_APP_TOKEN",
+  "PUSHOVER_USER_KEY",
+  "PUSHOVER_DEVICE",
+  "PUSHOVER_PRIORITY",
+  "PUSHOVER_SOUND",
+  "PICLAW_WEB_TLS_CERT",
+  "PICLAW_WEB_TLS_KEY",
+  "PICLAW_WEB_TOTP_SECRET",
+  "PICLAW_WEB_TOTP_WINDOW",
+  "PICLAW_WEB_SESSION_TTL",
+  "PICLAW_WEB_IDLE_TIMEOUT",
+  "PICLAW_WEB_INTERNAL_SECRET",
+  "PICLAW_WEB_WIDGET_TOKEN",
+  "PICLAW_WEB_PASSKEY_MODE",
+  "PICLAW_WEB_PUSH_SUBSCRIPTION_CAP",
+  "PICLAW_WEB_PUSH_VAPID_SUBJECT",
+  "PICLAW_WEB_TERMINAL_ENABLED",
+  "PICLAW_TERMINAL_IMAGE_PROTOCOL",
+  "PICLAW_WEB_COMPOSE_UPLOAD_LIMIT_MB",
+  "PICLAW_WEB_WORKSPACE_UPLOAD_LIMIT_MB",
+  "PICLAW_WEB_NOTIFICATION_DEBUG_LABELS",
+  "PICLAW_WEB_PERSIST_THINKING",
+  "PICLAW_WEB_PERSIST_THINKING_MAX_CHARS",
+  "PICLAW_WEB_VNC_ALLOW_DIRECT",
+  "PICLAW_VNC_ALLOW_DIRECT",
+  "PICLAW_WEB_VNC_TARGETS",
+  "PICLAW_VNC_TARGETS",
+  "PICLAW_DEBUG_CARD_SUBMISSIONS",
+  "PICLAW_TRUST_PROXY",
+  "PICLAW_SESSION_MAX_SIZE_MB",
+  "PICLAW_SESSION_AUTO_ROTATE",
+  "PICLAW_SESSION_ISOLATION",
+  "PICLAW_SESSION_FILE_PRELOAD_SANITIZE_MIN_BYTES",
+  "PICLAW_SESSION_TOOL_RESULT_MAX_PERSIST_BYTES",
+  "PICLAW_SESSION_TOOL_RESULT_PREVIEW_CHARS",
+  "PICLAW_RECOVERY_LOOP_GUARD_ENABLED",
+  "PICLAW_RECOVERY_LOOP_GUARD_MAX_FAILURES",
+  "PICLAW_RECOVERY_LOOP_GUARD_WINDOW_MS",
+  "PICLAW_TURN_AUTO_RECOVERY_ENABLED",
+  "PICLAW_TURN_AUTO_RECOVERY_MAX_ATTEMPTS",
+  "PICLAW_TURN_AUTO_RECOVERY_TOTAL_BUDGET_MS",
+  "PICLAW_STALE_PREFLIGHT_RECOVERY_MS",
+  "PICLAW_STALE_PREFLIGHT_BACKOFF_MS",
+  "PICLAW_PROGRESS_WATCHDOG_ENABLED",
+  "PICLAW_PROGRESS_WATCHDOG_TIMEOUT_MS",
+  "PICLAW_PROGRESS_WATCHDOG_RESTART_ON_STALL",
+  "PICLAW_PROGRESS_WATCHDOG_ESCALATE_ON_STALL",
+  "PICLAW_EXTERNAL_PROGRESS_WATCHDOG",
+  "PICLAW_TURN_MAX_TOOL_USE_MESSAGES",
+  "PICLAW_MID_TURN_TOOL_EXECUTION_HARD_CEILING",
+  "PICLAW_IDLE_AUTO_COMPACTION_DELAY_MS",
+  "PICLAW_PREPROMPT_COMPACTION_FOREGROUND_MS",
+  "PICLAW_AUTO_COMPACTION_ENABLED",
+  "PICLAW_SMART_COMPACTION_METHOD",
+  "PICLAW_REMOTE_COMPACTION_ENABLED",
+  "PICLAW_REMOTE_COMPACTION_TIMEOUT_MS",
+  "PICLAW_COMPACTION_TIMEOUT_MS",
+  "PICLAW_COMPACTION_BACKOFF_BASE_MS",
+  "PICLAW_COMPACTION_BACKOFF_MAX_MS",
+  "PICLAW_COMPACTION_THRESHOLD_PERCENT",
+  "PICLAW_COMPACTION_MAX_THRESHOLD_TOKENS",
+  "PICLAW_COMPACTION_BACKOFF_DECAY_FACTOR",
+  "PICLAW_SYSTEM_PROMPT_OVERHEAD_TOKENS",
+  "PICLAW_COMPACTION_REQUEST_OVERHEAD_TOKENS",
+  "PICLAW_TOKEN_ESTIMATE_SAFETY_MULTIPLIER",
+  "PICLAW_PROGRESSIVE_COMPACTION",
+  "PICLAW_SMART_COMPACTION_REASONING",
+  "PICLAW_AUTO_COMPACTION_SCOPE",
+  "PICLAW_COMPACTION_HARD_CEILING_PERCENT",
+  "PICLAW_COMPACTION_WARNING_THRESHOLD",
+  "PICLAW_TOOL_RESULT_COMPACTION_ENABLED",
+  "PICLAW_TOOL_RESULT_COMPACTION_TOOLS",
+  "PICLAW_TOOL_RESULT_SEMANTIC_SUMMARY_ENABLED",
+  "PICLAW_TOOL_RESULT_SEMANTIC_SUMMARY_MAX_INPUT_CHARS",
+  "PICLAW_TOOL_RESULT_SEMANTIC_SUMMARY_MAX_TOKENS",
+  "PICLAW_TOOL_RESULT_SEMANTIC_SUMMARY_TIMEOUT_MS",
+  "PICLAW_TOOL_OUTPUT_STORE_THRESHOLDS_BY_TOOL",
+  "PICLAW_TOOL_OUTPUT_STORE_BYTES",
+  "PICLAW_TOOL_OUTPUT_STORE_LINES",
+  "PICLAW_TOOL_OUTPUT_PREVIEW_LINES",
+  "PICLAW_TOOL_OUTPUT_PREVIEW_LINE_CHARS",
+  "PICLAW_WORKSPACE_SEARCH_ROOTS",
+  "PICLAW_INTERNAL_SECRET",
+  "PICLAW_REMOTE_INTEROP_ENABLED",
+  "PICLAW_REMOTE_INTEROP_ALLOW_HTTP",
+  "PICLAW_REMOTE_INTEROP_ALLOW_PRIVATE_NETWORK",
+  "PICLAW_REMOTE_INSTANCE_NAME",
+  "PICLAW_REMOTE_SHORT_CIRCUIT_ENABLED",
+  "PICLAW_REMOTE_INTEROP_DECISION_MODEL",
+  "PICLAW_WEB_EXTERNAL_URL",
+  "PICLAW_SCOPED_MODELS_ONLY",
+  "PICLAW_LOG_LEVEL",
+  "LOG_LEVEL",
+  "PICLAW_AGENT_LOG_RETENTION_MS",
+  "PICLAW_AGENT_LOG_RETENTION_DAYS",
+  "PICLAW_AGENT_LOG_CLEANUP_INTERVAL_MS",
+  "PICLAW_MAIN_SESSION_IDLE_TTL_MS",
+  "PICLAW_SIDE_SESSION_IDLE_TTL_MS",
+  "PICLAW_SESSION_IDLE_TTL_MS",
+  "PICLAW_SESSION_CLEANUP_INTERVAL_MS",
+  "PICLAW_MAIN_SESSION_POOL_MAX_SIZE",
+  "PICLAW_SESSION_POOL_MAX_SIZE",
+  "PICLAW_MAIN_SESSION_PRESSURE_RSS_BYTES",
+  "PICLAW_MAIN_SESSION_PRESSURE_IDLE_TTL_MS",
+  "PICLAW_MAIN_SESSION_PRESSURE_POOL_MAX_SIZE",
+  "PICLAW_TOOL_OUTPUT_RETENTION_MS",
+  "PICLAW_TOOL_OUTPUT_RETENTION_DAYS",
+  "PICLAW_TOOL_OUTPUT_CLEANUP_INTERVAL_MS",
+  "PICLAW_GITHUB_COPILOT_DYNAMIC_MODELS",
+  "PICLAW_GITHUB_COPILOT_MODELS_TIMEOUT_MS",
+  "PICLAW_MCP_TOOL_TIMEOUT_MS",
+  "PICLAW_PACKAGE_ROOT",
+  "PICLAW_UNKNOWN_MODEL_CONTEXT_WINDOW",
+  "PICLAW_WORKSPACE_SEARCH_EXTENSIONS",
+] as const;
+
+export function loadPiclawEnvConfig(): Record<string, string> {
+  return readEnvFile([...ENV_FILE_KEYS]);
+}
+
+export function nestedConfig(
+  config: Record<string, unknown>,
+  key: string,
+): Record<string, unknown> {
+  const value = config[key];
+  return value && typeof value === "object"
+    ? value as Record<string, unknown>
+    : config;
+}
